+++ src/database/index.ts
@@ -1,7 +1,7 @@
 import { ButtonInteraction, ModalSubmitInteraction, ThreadChannel } from "discord.js";
 import { TFunction } from "i18next";
 
-import { getEmbeds } from "../utils/parse_embeds";
+import { getEmbeds } from "../utils/parse";
 
 export async function getUserNameAndChar(interaction: ButtonInteraction | ModalSubmitInteraction, ul: TFunction<"translation", undefined>) {
 	const userEmbed = getEmbeds(ul, interaction?.message ?? undefined, "user");
